// ProyectoPortalEtan/src/app/interfaces/edocuenta.interface.ts
export default interface Edocuenta{
    id?: string;
    Matricula: string;
    MontoMensual: number;
    NumMensu: number;
    MensPag: number;
    MensPen: number;
    HrsVuelo: number;    
    HrsVueloPag: number;  
    HrsVueloPen: number;  
    HrsSimula: number;
    HrsSimulaPag: number;
    HrsSimulaPen: number;
}